rename projects to AzureDevOpsServerMCP
