+++ README.md
@@ -1,6 +1,6 @@
-# Azure DevOps MCP Fork
+# AzureDevOpsServerMCP
 
-This fork tracks the upstream Azure DevOps MCP project:
+AzureDevOpsServerMCP is a fork that tracks the upstream Azure DevOps MCP project:
 https://github.com/microsoft/azure-devops-mcp
 
 This fork is specifically intended to target Azure DevOps Server (on-prem) environments.
