--- dotnet/src/AzureDevOps.Mcp.Application/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace AzureDevOps.Mcp.Application;
-
-public class Class1
-{
-
-}
