--- dotnet/src/AzureDevOps.Mcp.Domain/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace AzureDevOps.Mcp.Domain;
-
-public class Class1
-{
-
-}
