--- dotnet/src/AzureDevOps.Mcp.Infrastructure.AzureDevOps/AzureDevOps.Mcp.Infrastructure.AzureDevOps.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="..\AzureDevOps.Mcp.Application\AzureDevOps.Mcp.Application.csproj" />
-    <ProjectReference Include="..\AzureDevOps.Mcp.Infrastructure.Configuration\AzureDevOps.Mcp.Infrastructure.Configuration.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.TeamFoundationServer.Client" />
-    <PackageReference Include="Microsoft.VisualStudio.Services.Client" />
-    <PackageReference Include="System.Configuration.ConfigurationManager" />
-  </ItemGroup>
-
-</Project>
