--- dotnet/src/AzureDevOps.Mcp.Infrastructure.AzureDevOps/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace AzureDevOps.Mcp.Infrastructure.AzureDevOps;
-
-public class Class1
-{
-
-}
