--- dotnet/src/AzureDevOps.Mcp.Infrastructure.AzureDevOps/WorkItemContextDto.cs
@@ -1,12 +1,0 @@
-using Microsoft.TeamFoundation.WorkItemTracking.WebApi.Models;
-
-namespace AzureDevOps.Mcp.Infrastructure.AzureDevOps;
-
-/// <summary>
-/// Data transfer object representing a work item with its context.
-/// </summary>
-public class WorkItemContextDto
-{
-    public required WorkItem WorkItem { get; set; }
-    public List<Comment> Comments { get; set; } = [];
-}
