--- dotnet/src/AzureDevOps.Mcp.Infrastructure.Configuration/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace AzureDevOps.Mcp.Infrastructure.Configuration;
-
-public class Class1
-{
-
-}
