--- dotnet/src/AzureDevOps.Mcp.Server/AzureDevOps.Mcp.Server.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <ItemGroup>
-    <ProjectReference Include="..\AzureDevOps.Mcp.Application\AzureDevOps.Mcp.Application.csproj" />
-    <ProjectReference Include="..\AzureDevOps.Mcp.Tools\AzureDevOps.Mcp.Tools.csproj" />
-    <ProjectReference Include="..\AzureDevOps.Mcp.Infrastructure.Configuration\AzureDevOps.Mcp.Infrastructure.Configuration.csproj" />
-    <ProjectReference Include="..\AzureDevOps.Mcp.Infrastructure.AzureDevOps\AzureDevOps.Mcp.Infrastructure.AzureDevOps.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="ModelContextProtocol.AspNetCore" />
-  </ItemGroup>
-
-</Project>
