--- dotnet/src/AzureDevOps.Mcp.Tools/AzureDevOps.Mcp.Tools.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="..\AzureDevOps.Mcp.Application\AzureDevOps.Mcp.Application.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="ModelContextProtocol" />
-  </ItemGroup>
-
-</Project>
