--- dotnet/src/AzureDevOps.Mcp.Tools/Class1.cs
@@ -1,13 +1,0 @@
-﻿using ModelContextProtocol.Server;
-
-namespace AzureDevOps.Mcp.Tools;
-
-[McpServerToolType]
-public static class HealthTools
-{
-	[McpServerTool]
-	public static string Ping()
-	{
-		return "pong";
-	}
-}
