+++ dotnet/src/AzureDevOpsServerMCP.Application/AzureDevOpsServerMCP.Application.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\AzureDevOpsServerMCP.Domain\AzureDevOpsServerMCP.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
