+++ dotnet/src/AzureDevOpsServerMCP.Application/Class1.cs
@@ -1,0 +1,6 @@
+namespace AzureDevOpsServerMCP.Application;
+
+public class Class1
+{
+
+}
