+++ dotnet/src/AzureDevOpsServerMCP.Domain/AzureDevOpsServerMCP.Domain.csproj
