+++ dotnet/src/AzureDevOpsServerMCP.Domain/Class1.cs
@@ -1,0 +1,6 @@
+namespace AzureDevOpsServerMCP.Domain;
+
+public class Class1
+{
+
+}
