+++ dotnet/src/AzureDevOpsServerMCP.Infrastructure.AzureDevOps/AzureDevOpsServerMCP.Infrastructure.AzureDevOps.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\AzureDevOpsServerMCP.Application\AzureDevOpsServerMCP.Application.csproj" />
+    <ProjectReference Include="..\AzureDevOpsServerMCP.Infrastructure.Configuration\AzureDevOpsServerMCP.Infrastructure.Configuration.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.TeamFoundationServer.Client" />
+    <PackageReference Include="Microsoft.VisualStudio.Services.Client" />
+    <PackageReference Include="System.Configuration.ConfigurationManager" />
+  </ItemGroup>
+
+</Project>
