+++ dotnet/src/AzureDevOpsServerMCP.Infrastructure.AzureDevOps/Class1.cs
@@ -1,0 +1,6 @@
+namespace AzureDevOpsServerMCP.Infrastructure.AzureDevOps;
+
+public class Class1
+{
+
+}
