+++ dotnet/src/AzureDevOpsServerMCP.Infrastructure.Configuration/AzureDevOpsServerMCP.Infrastructure.Configuration.csproj
