+++ dotnet/src/AzureDevOpsServerMCP.Infrastructure.Configuration/Class1.cs
@@ -1,0 +1,6 @@
+namespace AzureDevOpsServerMCP.Infrastructure.Configuration;
+
+public class Class1
+{
+
+}
