+++ dotnet/src/AzureDevOpsServerMCP.Server/AzureDevOpsServerMCP.Server.csproj.user
