+++ dotnet/src/AzureDevOpsServerMCP.Server/Properties/launchSettings.json
