+++ dotnet/src/AzureDevOpsServerMCP.Server/appsettings.json
