+++ dotnet/src/AzureDevOpsServerMCP.Tools/AzureDevOpsServerMCP.Tools.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\AzureDevOpsServerMCP.Application\AzureDevOpsServerMCP.Application.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="ModelContextProtocol" />
+  </ItemGroup>
+
+</Project>
