+++ dotnet/src/AzureDevOpsServerMCP.Tools/Class1.cs
@@ -1,0 +1,13 @@
+using ModelContextProtocol.Server;
+
+namespace AzureDevOpsServerMCP.Tools;
+
+[McpServerToolType]
+public static class HealthTools
+{
+	[McpServerTool]
+	public static string Ping()
+	{
+		return "pong";
+	}
+}
