--- dotnet/tests/AzureDevOps.Mcp.IntegrationTests/AzureDevOps.Mcp.IntegrationTests.csproj
@@ -1,25 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="coverlet.collector" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="xunit" />
-    <PackageReference Include="xunit.runner.visualstudio" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Using Include="Xunit" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\AzureDevOps.Mcp.Application\AzureDevOps.Mcp.Application.csproj" />
-    <ProjectReference Include="..\..\src\AzureDevOps.Mcp.Tools\AzureDevOps.Mcp.Tools.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="Fixtures\work-item-context-result.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-</Project>
