--- dotnet/tests/AzureDevOps.Mcp.UnitTests/AzureDevOps.Mcp.UnitTests.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="coverlet.collector" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="xunit" />
-    <PackageReference Include="xunit.runner.visualstudio" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\AzureDevOps.Mcp.Application\AzureDevOps.Mcp.Application.csproj" />
-    <ProjectReference Include="..\..\src\AzureDevOps.Mcp.Infrastructure.AzureDevOps\AzureDevOps.Mcp.Infrastructure.AzureDevOps.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Using Include="Xunit" />
-  </ItemGroup>
-
-</Project>
